requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
dj-database-url==1.2.0
Django==4.1.7
django-filter==22.1
django-heroku==0.3.1
django-jazzmin==2.6.0
django-shopping-cart==0.1
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
